server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Kim Minseok",email = "minseok@chungbuk.ac.kr"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dramatiq[redis] (>=2.0.0,<3.0.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "google-cloud-texttospeech (>=2.33.0,<3.0.0)",
    "minio (>=7.2.20,<8.0.0)",
    "mutagen (>=1.47.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "server", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
